requirements.txt
Django==2.1.5
django-environ==0.4.5
djangorestframework==3.9.1
mysqlclient==1.4.0
pip==18.1
pytz==2018.9
virtualenv==16.0.0
setuptools==40.6.3
wheel==0.32.3
